Improved patch and changed documentation
Instead of removing items, they will be disabled by default. This way the user can enable it later, if desired.
Some default JVM arguments that either impacted performance, increased RAM usage or was not being recognized were removed.
`-d64` flag removed from list as it was not recognized.

This patch update is meant to create a `Essentials Lite` version of the newest `1.8.1 - Light` one.
Need more improvements to be ready yet.